# Hardware performance counter support (via `rdpmc`).