--- a/measureme/src/counters.rs
+++ b/measureme/src/counters.rs
@@ -11,9 +11,8 @@ macro_rules! really_warn {
 }
 
 pub enum Counter {
-    Zero(InstructionsMinusIrqs),
-    WallTime(InstructionsMinusIrqs, WallTime),
-    Instructions(InstructionsMinusIrqs, Instructions),
+    WallTime(WallTime),
+    Instructions(Instructions),
     InstructionsMinusIrqs(InstructionsMinusIrqs),
     InstructionsMinusRaw0420(InstructionsMinusRaw0420),
 }
@@ -21,17 +20,12 @@ pub enum Counter {
 impl Counter {
     pub fn by_name(name: &str) -> Result<Self, Box<dyn Error + Send + Sync>> {
         Ok(match name {
-            "0" => Counter::Zero(InstructionsMinusIrqs::new()?),
-            "t" | WallTime::NAME => {
-                Counter::WallTime(InstructionsMinusIrqs::new()?, WallTime::new())
-            }
-            "i" | Instructions::NAME => {
-                Counter::Instructions(InstructionsMinusIrqs::new()?, Instructions::new()?)
-            }
-            "I" | InstructionsMinusIrqs::NAME => {
+            WallTime::NAME => Counter::WallTime(WallTime::new()),
+            Instructions::NAME => Counter::Instructions(Instructions::new()?),
+            InstructionsMinusIrqs::NAME => {
                 Counter::InstructionsMinusIrqs(InstructionsMinusIrqs::new()?)
             }
-            "r" | InstructionsMinusRaw0420::NAME => {
+            InstructionsMinusRaw0420::NAME => {
                 Counter::InstructionsMinusRaw0420(InstructionsMinusRaw0420::new()?)
             }
             _ => return Err(format!("{:?} is not a valid counter name", name).into()),
@@ -40,12 +34,11 @@ impl Counter {
 
     pub(super) fn describe_as_json(&self) -> String {
         let (name, units) = match self {
-            Counter::Zero(_) => ("zero", "[]"),
-            Counter::WallTime(..) => (
+            Counter::WallTime(_) => (
                 WallTime::NAME,
                 r#"[["ns", 1], ["μs", 1000], ["ms", 1000000], ["s", 1000000000]]"#,
             ),
-            Counter::Instructions(..) => (Instructions::NAME, r#"[["instructions", 1]]"#),
+            Counter::Instructions(_) => (Instructions::NAME, r#"[["instructions", 1]]"#),
             Counter::InstructionsMinusIrqs(_) => {
                 (InstructionsMinusIrqs::NAME, r#"[["instructions", 1]]"#)
             }
@@ -59,9 +52,8 @@ impl Counter {
     #[inline]
     pub(super) fn since_start(&self) -> u64 {
         match self {
-            Counter::Zero(_) => 0,
-            Counter::WallTime(_, counter) => counter.since_start(),
-            Counter::Instructions(_, counter) => counter.since_start(),
+            Counter::WallTime(counter) => counter.since_start(),
+            Counter::Instructions(counter) => counter.since_start(),
             Counter::InstructionsMinusIrqs(counter) => counter.since_start(),
             Counter::InstructionsMinusRaw0420(counter) => counter.since_start(),
         }
@@ -137,17 +129,6 @@ impl InstructionsMinusIrqs {
     fn since_start(&self) -> u64 {
         let (instructions, irqs) = (&self.instructions, &self.irqs).read();
         instructions.wrapping_sub(irqs).wrapping_sub(self.start)
-    }
-}
-
-// HACK(eddyb) dump total `instructions-minus-irqs:u` for statistics.
-impl Drop for InstructionsMinusIrqs {
-    fn drop(&mut self) {
-        eprintln!(
-            "pid={:06} instructions-minus-irqs:u={}",
-            std::process::id(),
-            self.since_start(),
-        );
     }
 }