@ambiso 2020-11-04
Hi,

I'm not sure if this is directly relevant to this PR,
and I very well may not be telling anyone anything new.

Still, since I saw the insistence to remove all noise from the measurements I just wanted to make sure, that people are aware of research that suggests many optimizations performed by LLVM on `-O3` are for naught when tested with random stack and binary layout [1,2,3].
This would suggest that you would actually not want to simply disable ASLR, but test on every possible layout.

If this is the wrong pull request to post this comment, or you are aware of this research, or that it doesn't matter for what you're trying to achieve, please feel free to ignore this comment, move it, or whatever.

Take care,
ambiso

[1]: https://users.cs.northwestern.edu/~robby/courses/322-2013-spring/mytkowicz-wrong-data.pdf
[2]: https://youtu.be/r-TLSBdHe1A?t=948
[3]: https://github.com/ccurtsinger/stabilizer