@eddyb 2020-11-04
@ambiso The way ASLR feeds into this is roughly `HashMap<*const T, ...>` (the equivalent of `unordered_map<T*, ...>` in C++), i.e. hashing pointer addresses, meaning the choice of addresses affects the order in the hash map, and therefore the performance characteristics, by somewhere around `±0.01%`.

I'm not sure how optimizations, or *stack layout* feeds into it, but we're not (yet) measuring anything like cache effectiveness, the entire problem with "ASLR non-determinism" in our case is our fault for relying on (interned) pointer addresses for cheaper hashing (and avoiding having to *store* a hash somewhere at all).